+++ packages/picasso-rich-text-editor/src/plugins/ImagePlugin/ImagePlugin.tsx
@@ -1,23 +1,12 @@
-import React, { useEffect } from 'react'
-import { useModal } from '@toptal/picasso/utils'
-import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext'
-import {
-  $createParagraphNode,
-  $getSelection,
-  $insertNodes,
-  $isRootOrShadowRoot,
-  COMMAND_PRIORITY_EDITOR,
-} from 'lexical'
-import { $wrapNodeInElement } from '@lexical/utils'
+import React from 'react'
 
 import type { RTEPlugin } from '../api'
 import { RTEPluginMeta, Toolbar } from '../api'
 import { ImagePluginButton, ImagePluginModal } from './components'
-import { INSERT_IMAGE_COMMAND } from './commands'
-import type { ImageNodePayload } from './nodes/ImageNode'
-import { $createImageNode, ImageNode } from './nodes/ImageNode'
-import type { OnUploadCallback, UploadedImage } from './types'
+import { ImageNode } from './nodes/ImageNode'
+import type { OnUploadCallback } from './types'
 import type { ImagePluginModalProps } from './components/ImagePluginModal'
+import { useImagePlugin } from './hooks'
 
 const PLUGIN_NAME = 'image'
 
@@ -34,45 +23,7 @@ const ImagePlugin: RTEPlugin = ({
   onUpload,
   'data-testid': testId,
 }: Props) => {
-  const { isOpen, hideModal, showModal } = useModal()
-  const [editor] = useLexicalComposerContext()
-
-  useEffect(() => {
-    return editor.registerCommand(
-      INSERT_IMAGE_COMMAND,
-      (imagePayload: ImageNodePayload) => {
-        const imageNode = $createImageNode(imagePayload)
-
-        $insertNodes([imageNode])
-        if ($isRootOrShadowRoot(imageNode.getParentOrThrow())) {
-          $wrapNodeInElement(imageNode, $createParagraphNode).selectEnd()
-        }
-
-        return true
-      },
-      COMMAND_PRIORITY_EDITOR
-    )
-  }, [editor])
-
-  const onSubmit: ImagePluginModalProps['onSubmit'] = (
-    image: UploadedImage,
-    altText: string
-  ) => {
-    editor.update(() => {
-      if (image.url) {
-        const imageContainer = $createParagraphNode()
-        const imageNode = $createImageNode({
-          alt: altText,
-          src: image.url,
-        })
-
-        imageContainer.append(imageNode)
-        const selection = $getSelection()
-
-        selection?.insertNodes([imageContainer])
-      }
-    })
-  }
+  const { modalIsOpen, hideModal, showModal, onSubmit } = useImagePlugin()
 
   return (
     <>
@@ -82,7 +33,7 @@ const ImagePlugin: RTEPlugin = ({
       <ImagePluginModal
         accept={accept}
         maxSize={maxSize}
-        isOpen={isOpen}
+        isOpen={modalIsOpen}
         onUpload={onUpload}
         onSubmit={onSubmit}
         onClose={hideModal}
