+++ packages/picasso-rich-text-editor/src/plugins/ImagePlugin/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useImageUploader } from './use-image-uploader'
+export { useImagePlugin } from './use-image-plugin'
